Explicit unboxing
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/pool/trunk@1021900 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/pool/impl/StackKeyedObjectPool.java
@@ -417,7 +417,7 @@ public class StackKeyedObjectPool<K,V> extends BaseKeyedObjectPool<K,V> implemen
      */
     private int getActiveCount(K key) {
         try {
-            return _activeCount.get(key);
+            return _activeCount.get(key).intValue();
         } catch(NoSuchElementException e) {
             return 0;
         } catch(NullPointerException e) {
